backend/requirements.txt
# FastAPI and dependencies
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.6.0
pydantic-settings>=2.1.0

# Database related
sqlalchemy>=2.0.0
alembic>=1.13.0
pymysql>=1.1.0
cryptography>=42.0.0

# Logging and monitoring
elasticsearch>=8.0.0
loguru>=0.7.0
prometheus-client>=0.19.0
opentelemetry-api>=1.23.0
opentelemetry-sdk>=1.23.0
opentelemetry-instrumentation-fastapi>=0.44b0

# Utils
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0
pandas>=2.2.0
numpy>=1.26.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
locust>=2.21.0 